Fixed Package Renaming
1.) Must: Change the pom main target, change the package paths in every
class. Make sure the FXML is referencing the correct java file.
