+++ code/pom.xml
@@ -4,7 +4,7 @@
 	
 	<!-- project info -->
 	<groupId>edu.westga</groupId>
-	<artifactId>Sample-JavaFX-Project</artifactId>
+	<artifactId>Budget-Software</artifactId>
 	<version>1.0-SNAPSHOT</version>
 	
 	<properties>
@@ -55,7 +55,7 @@
 				<artifactId>javafx-maven-plugin</artifactId>
 				<version>0.0.6</version>
 				<configuration>
-					<mainClass>edu.westga.comp4420.javafx_sample.Main</mainClass>
+					<mainClass>edu.westga.comp4420.budget_software.Main</mainClass>
 				</configuration>
 			</plugin>
 			
@@ -113,8 +113,8 @@
 								<rule>
 									<element>PACKAGE</element>
 									<excludes>
-										<exclude>edu.westga.comp4420.lab1.view.codebehind</exclude>
-										<exclude>edu.westga.comp4420.lab1</exclude>
+										<exclude>edu.westga.comp4420.budget_software.view.codebehind</exclude>
+										<exclude>edu.westga.comp4420.budget_software</exclude>
 									</excludes>
 									<limits>
 										<limit>
