+++ code/src/main/java/edu/westga/comp4420/budget_software/view/codebehind/MainWindow.java
@@ -1,4 +1,4 @@
-package edu.westga.comp4420.javafx_sample.view.codebehind;
+package edu.westga.comp4420.budget_software.view.codebehind;
 
 
 /**
